Creating blank method to BTN-Interest

+++ src/MathOperations.java
@@ -1,10 +1,10 @@
 import java.math.BigDecimal;
 
 public class MathOperations {
-	public static double truncate(double number, int decimal_digits){
+	public static float truncate(double number, int decimal_digits){
 		
 		BigDecimal bd = BigDecimal.valueOf(number);
 		bd = bd.setScale(decimal_digits, BigDecimal.ROUND_DOWN);
-		return bd.doubleValue();
+		return bd.floatValue();
 	}
 }
